changed unique access for a job

+++ src/job/job.resolver.ts
@@ -4,8 +4,8 @@ import {
   Mutation,
   Args,
   ID,
-  // ResolveField,
-  // Parent,
+  ResolveField,
+  Parent,
 } from '@nestjs/graphql';
 import { JobService } from './job.service';
 
@@ -13,6 +13,8 @@ import { CreateJobInput } from './dto/create-job.input';
 import { UpdateJobInput } from './dto/update-job.input';
 import { Job } from './job.model';
 import { PrismaService } from 'src/prisma/prisma.service';
+
+import { Profile } from 'src/profile/profile.model';
 import { BadRequestException } from '@nestjs/common';
 
 @Resolver(() => Job)
@@ -24,15 +26,15 @@ export class JobResolver {
 
   @Mutation(() => Job)
   async createJob(@Args('data') createJobInput: CreateJobInput) {
-    // const check = await this.prisma.profile.findUnique({
-    //   where: {
-    //     id: createJobInput.profileId,
-    //   },
-    //   select: { type: true },
-    // });
-    // if (!check) {
-    //   return new BadRequestException({ error: 'Recruiter can only post jobs' });
-    // }
+    const check = await this.prisma.profile.findUnique({
+      where: {
+        id: createJobInput.profileId,
+      },
+      select: { recruiter: true },
+    });
+    if (!check) {
+      return new BadRequestException({ error: 'Recruiter can only post jobs' });
+    }
     return await this.jobService.create(createJobInput);
   }
 
@@ -42,8 +44,8 @@ export class JobResolver {
   }
 
   @Query(() => Job, { name: 'job' })
-  findOne(@Args('id', { type: () => ID }) id: string) {
-    return this.jobService.findOne(id);
+  findJobOne(@Args('id', { type: () => ID }) id: string) {
+    return this.jobService.findJobOne(id);
   }
 
   @Mutation(() => Job)
@@ -51,43 +53,43 @@ export class JobResolver {
     @Args('id', { type: () => ID }) id: string,
     @Args('data') updateJobInput: UpdateJobInput,
   ) {
-    // const check = await this.prisma.profile.findUnique({
-    //   where: {
-    //     id: updateJobInput.profileId,
-    //   },
-    //   select: { type: true },
-    // });
-    // if (!check) {
-    //   return new BadRequestException({
-    //     error: 'Recruiter can only update jobs',
-    //   });
-    // }
-    await this.jobService.findOne(id);
+    const check = await this.prisma.profile.findUnique({
+      where: {
+        id: updateJobInput.profileId,
+      },
+      select: { recruiter: true },
+    });
+    if (!check) {
+      return new BadRequestException({
+        error: 'Recruiter can only update jobs',
+      });
+    }
+    await this.jobService.findJobOne(id);
     return this.jobService.update(id, updateJobInput);
   }
 
   @Mutation(() => Job)
   async removeJob(@Args('id', { type: () => ID }) id: string) {
-    // const check = await this.prisma.profile.findUnique({
-    //   where: {
-    //     id,
-    //   },
-    //   select: { type: true },
-    // });
-    // if (!check) {
-    //   return new BadRequestException({
-    //     error: 'Recruiter can only delete jobs',
-    //   });
-    // }
-    await this.jobService.findOne(id);
+    const check = await this.prisma.profile.findUnique({
+      where: {
+        id,
+      },
+      select: { recruiter: true },
+    });
+    if (!check) {
+      return new BadRequestException({
+        error: 'Recruiter can only delete jobs',
+      });
+    }
+    await this.jobService.findJobOne(id);
     return this.jobService.remove(id);
   }
-  // @ResolveField(() => Profile)
-  // async job(@Parent() job: Job) {
-  //   return await this.prisma.job
-  //     .findUnique({ where: { id: job.id } })
-  //     .profile();
-  // }
+  @ResolveField(() => Profile)
+  async profile(@Parent() job: Job) {
+    return await this.prisma.job
+      .findUnique({ where: { id: job.id } })
+      .profile();
+  }
 
   // @ResolveField(() => [JobApplication])
   // async jobApplication(@Parent() job: Job) {
